# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "CurrentReadingDto": {
        "additionalProperties": false,
        "properties": {
          "baromAbsIn": {
            "format": "double",
            "nullable": true,
            "type": "number"
          },
          "baromRelIn": {
            "format": "double",
            "nullable": true,
            "type": "number"
          },
          "battOut": {
            "format": "int32",
            "nullable": true,
            "type": "integer"
          },
          "dailyHighTempF": {
            "format": "double",
            "nullable": true,
            "type": "number"
          },
          "dailyLowTempF": {
            "format": "double",
            "nullable": true,
            "type": "number"
          },
          "dailyRainIn": {
            "format": "double",
            "nullable": true,
            "type": "number"
          },
          "deviceId": {
            "nullable": true,
            "type": "string"
          },
          "deviceName": {
            "nullable": true,
            "type": "string"
          },
          "dewPoint": {
            "format": "double",
            "nullable": true,
            "type": "number"
          },
          "dewPointIn": {
            "format": "double",
            "nullable": true,
            "type": "number"
          },
          "eventRainIn": {
            "format": "double",
            "nullable": true,
            "type": "number"
          },
          "feelsLike": {
            "format": "double",
            "nullable": true,
            "type": "number"
          },
          "feelsLikeIn": {
            "format": "double",
            "nullable": true,
            "type": "number"
          },
          "hourlyRainIn": {
            "format": "double",
            "nullable": true,
            "type": "number"
          },
          "humidity": {
            "format": "int32",
            "nullable": true,
            "type": "integer"
          },
          "humidityIn": {
            "format": "int32",
            "nullable": true,
            "type": "integer"
          },
          "lastRain": {
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "maxDailyGust": {
            "format": "double",
            "nullable": true,
            "type": "number"
          },
          "monthlyRainIn": {
            "format": "double",
            "nullable": true,
            "type": "number"
          },
          "nwsPresentWeather": {
            "nullable": true,
            "type": "string"
          },
          "nwsRawMetar": {
            "nullable": true,
            "type": "string"
          },
          "nwsSkyConditions": {
            "nullable": true,
            "type": "string"
          },
          "nwsTextDescription": {
            "nullable": true,
            "type": "string"
          },
          "omCloudCover": {
            "format": "int32",
            "nullable": true,
            "type": "integer"
          },
          "omPrecipProbability": {
            "format": "int32",
            "nullable": true,
            "type": "integer"
          },
          "omPrecipSumIn": {
            "format": "double",
            "nullable": true,
            "type": "number"
          },
          "omSunrise": {
            "nullable": true,
            "type": "string"
          },
          "omSunset": {
            "nullable": true,
            "type": "string"
          },
          "omUvIndexMax": {
            "format": "int32",
            "nullable": true,
            "type": "integer"
          },
          "omWeatherDescription": {
            "nullable": true,
            "type": "string"
          },
          "omWindDirDominant": {
            "format": "int32",
            "nullable": true,
            "type": "integer"
          },
          "omWindGustMax": {
            "format": "double",
            "nullable": true,
            "type": "number"
          },
          "omWindSpeedMax": {
            "format": "double",
            "nullable": true,
            "type": "number"
          },
          "receivedAtUtc": {
            "format": "date-time",
            "type": "string"
          },
          "solarRadiation": {
            "format": "double",
            "nullable": true,
            "type": "number"
          },
          "source": {
            "nullable": true,
            "type": "string"
          },
          "tempF": {
            "format": "double",
            "nullable": true,
            "type": "number"
          },
          "tempInF": {
            "format": "double",
            "nullable": true,
            "type": "number"
          },
          "timestampUtc": {
            "format": "date-time",
            "type": "string"
          },
          "totalRainIn": {
            "format": "double",
            "nullable": true,
            "type": "number"
          },
          "tz": {
            "nullable": true,
            "type": "string"
          },
          "uv": {
            "format": "int32",
            "nullable": true,
            "type": "integer"
          },
          "weeklyRainIn": {
            "format": "double",
            "nullable": true,
            "type": "number"
          },
          "windDir": {
            "format": "int32",
            "nullable": true,
            "type": "integer"
          },
          "windGustMph": {
            "format": "double",
            "nullable": true,
            "type": "number"
          },
          "windSpeedMph": {
            "format": "double",
            "nullable": true,
            "type": "number"
          },
          "yearlyRainIn": {
            "format": "double",
            "nullable": true,
            "type": "number"
          }
        },
        "required": [
          "deviceId"
        ],
        "type": "object"
      },
...
      "NeighborConfigDto": {
        "additionalProperties": false,
        "properties": {
          "ambientOpenMaxRadiusMiles": {
            "format": "double",
            "type": "number"
          },
          "comparisonRadiusMiles": {
            "format": "double",
            "type": "number"
          },
          "discoveryLocationQuery": {
            "nullable": true,
            "type": "string"
          },
          "enabledProviders": {
            "items": {
              "type": "string"
            },
            "nullable": true,
            "type": "array"
          },
          "enabledStationMacAddresses": {
            "items": {
              "type": "string"
            },
            "nullable": true,
            "type": "array"
          },
          "isAmbientOpenAvailable": {
            "type": "boolean"
          },
          "isEnabled": {
            "type": "boolean"
          },
          "maxAgeMinutes": {
            "format": "int32",
            "type": "integer"
          },
          "minStations": {
            "format": "int32",
            "type": "integer"
          },
          "municipality": {
            "nullable": true,
            "type": "string"
          },
          "pinnedStations": {
            "items": {
              "$ref": "#/components/schemas/PinnedNeighborStationDto"
            },
            "nullable": true,
            "type": "array"
          },
          "radiusMiles": {
            "format": "double",
            "type": "number"
          },
          "refreshIntervalMinutes": {
            "format": "int32",
            "type": "integer"
          }
        },
        "required": [
          "comparisonRadiusMiles",
          "enabledProviders",
          "isEnabled",
          "maxAgeMinutes",
          "minStations",
          "radiusMiles",
          "refreshIntervalMinutes"
        ],
        "type": "object"
      },
      "NeighborStationDto": {
        "additionalProperties": false,
        "properties": {
          "baromAbsIn": {
            "format": "double",
            "nullable": true,
            "type": "number"
          },
          "baromRelIn": {
            "format": "double",
            "nullable": true,
            "type": "number"
          },
          "dailyRainIn": {
            "format": "double",
            "nullable": true,
            "type": "number"
          },
          "dewPoint": {
            "format": "double",
            "nullable": true,
            "type": "number"
          },
          "discoveryKind": {
            "nullable": true,
            "type": "string"
          },
          "distanceMiles": {
            "format": "double",
            "type": "number"
          },
          "feelsLike": {
            "format": "double",
            "nullable": true,
            "type": "number"
          },
          "freshnessMinutes": {
            "format": "int32",
            "nullable": true,
            "type": "integer"
          },
          "hourlyRainIn": {
            "format": "double",
            "nullable": true,
            "type": "number"
          },
          "humidity": {
            "format": "int32",
            "nullable": true,
            "type": "integer"
          },
          "lastObservedAtUtc": {
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "lat": {
            "format": "double",
            "type": "number"
          },
          "lon": {
            "format": "double",
            "type": "number"
          },
          "monthlyRainIn": {
            "format": "double",
            "nullable": true,
            "type": "number"
          },
          "name": {
            "nullable": true,
            "type": "string"
          },
          "provider": {
            "nullable": true,
            "type": "string"
          },
          "solarRadiation": {
            "format": "double",
            "nullable": true,
            "type": "number"
          },
          "sourceId": {
            "nullable": true,
            "type": "string"
          },
          "tempF": {
            "format": "double",
            "nullable": true,
            "type": "number"
          },
          "uv": {
            "format": "int32",
            "nullable": true,
            "type": "integer"
          },
          "weeklyRainIn": {
            "format": "double",
            "nullable": true,
            "type": "number"
          },
          "windDir": {
            "format": "int32",
            "nullable": true,
            "type": "integer"
          },
          "windGustMph": {
            "format": "double",
            "nullable": true,
            "type": "number"
          },
          "windSpeedMph": {
            "format": "double",
            "nullable": true,
            "type": "number"
          },
          "yearlyRainIn": {
            "format": "double",
            "nullable": true,
            "type": "number"
          }
        },
        "required": [
          "distanceMiles",
          "lat",
          "lon",
          "provider",
          "sourceId"
        ],
        "type": "object"
      },
...
      "PinnedNeighborStation": {
        "additionalProperties": false,
        "properties": {
          "displayLabel": {
            "nullable": true,
            "type": "string"
          },
          "isEnabled": {
            "type": "boolean"
          },
          "provider": {
            "nullable": true,
            "type": "string"
          },
          "selectedMetricKeys": {
            "items": {
              "type": "string"
            },
            "nullable": true,
            "type": "array"
          },
          "sourceId": {
            "nullable": true,
            "type": "string"
          }
        },
        "type": "object"
      },
      "PinnedNeighborStationDto": {
        "additionalProperties": false,
        "properties": {
          "displayLabel": {
            "nullable": true,
            "type": "string"
          },
          "isEnabled": {
            "type": "boolean"
          },
          "provider": {
            "nullable": true,
            "type": "string"
          },
          "selectedMetricKeys": {
            "items": {
              "type": "string"
            },
            "nullable": true,
            "type": "array"
          },
          "sourceId": {
            "nullable": true,
            "type": "string"
          }
        },
        "required": [
          "provider",
          "sourceId"
        ],
        "type": "object"
      },
      "ProblemDetails": {
        "additionalProperties": {},
        "properties": {
          "detail": {
            "nullable": true,
            "type": "string"
          },
          "instance": {
            "nullable": true,
            "type": "string"
          },
          "status": {
            "format": "int32",
            "nullable": true,
            "type": "integer"
          },
          "title": {
            "nullable": true,
            "type": "string"
          },
          "type": {
            "nullable": true,
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "UpdateNeighborConfigCommand": {
        "additionalProperties": false,
        "properties": {
          "comparisonRadiusMiles": {
            "format": "double",
            "nullable": true,
            "type": "number"
          },
          "discoveryLocationQuery": {
            "nullable": true,
            "type": "string"
          },
          "enabledProviders": {
            "items": {
              "type": "string"
            },
            "nullable": true,
            "type": "array"
          },
          "enabledStationMacAddresses": {
            "items": {
              "type": "string"
            },
            "nullable": true,
            "type": "array"
          },
          "isEnabled": {
            "type": "boolean"
          },
          "maxAgeMinutes": {
            "format": "int32",
            "type": "integer"
          },
          "minStations": {
            "format": "int32",
            "type": "integer"
          },
          "municipality": {
            "nullable": true,
            "type": "string"
          },
          "pinnedStations": {
            "items": {
              "$ref": "#/components/schemas/PinnedNeighborStation"
            },
            "nullable": true,
            "type": "array"
          },
          "radiusMiles": {
            "format": "double",
            "type": "number"
          },
          "refreshIntervalMinutes": {
            "format": "int32",
            "type": "integer"
          }
        },
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "bearerFormat": "JWT",
        "description": "Paste your Auth0 access token (without \u0027Bearer \u0027 prefix).",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "BFF API for the Ambient Weather Dashboard web application.",
    "title": "Ambient Weather Dashboard API",
    "version": "v1"
  },
  "openapi": "3.0.4",
  "paths": {
...
    "/api/neighbors/comparison-stations": {
      "get": {
        "parameters": [
          {
            "in": "query",
            "name": "macAddress",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/NeighborStationDto"
                  },
                  "type": "array"
                }
              },
              "text/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/NeighborStationDto"
                  },
                  "type": "array"
                }
              },
              "text/plain": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/NeighborStationDto"
                  },
                  "type": "array"
                }
              }
            },
            "description": "List of contributing neighbor stations."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not authenticated."
          }
        },
        "security": [
          {}
        ],
        "summary": "Returns the Ambient Weather stations contributing to neighbor comparison for an owned station.",
        "tags": [
          "Neighbors"
        ]
      }
    },
    "/api/neighbors/config": {
      "get": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NeighborConfigDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/NeighborConfigDto"
                }
              },
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/NeighborConfigDto"
                }
              }
            },
            "description": "Neighbor configuration."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not authenticated."
          }
        },
        "security": [
          {}
        ],
        "summary": "Returns the neighbor comparison configuration for the authenticated user,\nseeding defaults on first access. Includes \u0060isAmbientOpenAvailable\u0060 so\nthe frontend can enable or disable the provider checkbox accordingly.",
        "tags": [
          "Neighbors"
        ]
      },
      "put": {
        "requestBody": {
          "content": {
            "application/*\u002Bjson": {
              "schema": {
                "$ref": "#/components/schemas/UpdateNeighborConfigCommand"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateNeighborConfigCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateNeighborConfigCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NeighborConfigDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/NeighborConfigDto"
                }
              },
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/NeighborConfigDto"
                }
              }
            },
            "description": "Updated neighbor configuration."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Validation error."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not authenticated."
          }
        },
        "security": [
          {}
        ],
        "summary": "Updates the neighbor comparison configuration for the authenticated user.",
        "tags": [
          "Neighbors"
        ]
      }
    },
    "/api/neighbors/refresh": {
      "post": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/NeighborStationDto"
                  },
                  "type": "array"
                }
              },
              "text/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/NeighborStationDto"
                  },
                  "type": "array"
                }
              },
              "text/plain": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/NeighborStationDto"
                  },
                  "type": "array"
                }
              }
            },
            "description": "OK"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unauthorized"
          },
          "428": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Precondition Required"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Too Many Requests"
          }
        },
        "security": [
          {}
        ],
        "summary": "Triggers a neighbor station discovery refresh and returns the updated station list.",
        "tags": [
          "Neighbors"
        ]
      }
    },
    "/api/neighbors/stations/current": {
      "get": {
        "parameters": [
          {
            "in": "query",
            "name": "provider",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "sourceId",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrentReadingDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrentReadingDto"
                }
              },
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CurrentReadingDto"
                }
              }
            },
            "description": "Current reading for the pinned station."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not authenticated."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Station not found in cache \u2014 refresh neighbor stations first."
          },
          "428": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Neighbor comparison is disabled."
          },
          "429": {
            "description": "Rate limit exceeded."
          }
        },
        "security": [
          {}
        ],
        "summary": "Clears the neighbor station cache and triggers an immediate re-discovery.\nReturns the refreshed station list.",
        "tags": [
          "Neighbors"
        ]
      }
    },
...
  },
...
}